+++ test/controllers/api/v1/projects_controller_test.rb
@@ -9,12 +9,30 @@ class Api::V1::ProjectsControllerTest < ActionDispatch::IntegrationTest
   end
 
   test "should get show" do
-    get api_v1_project_path(authentication_token: @project.id_and_token, format: "json")
+    get api_v1_project_path(
+      authentication_token: @project.id_and_token,
+      format: "json"
+    )
     assert_response :success
   end
 
   test "should get survey info" do
-    get api_v1_survey_info_path(authentication_token: @project.id_and_token, event: events(:api_event), design: designs(:api_design), format: "json")
+    get api_v1_survey_info_path(
+      authentication_token: @project.id_and_token,
+      event: events(:api_event),
+      design: designs(:api_design),
+      format: "json"
+    )
+    assert_response :success
+  end
+
+  test "should get subject counts" do
+    get api_v1_subject_counts_path(
+      authentication_token: @project.id_and_token,
+      expressions: ["api_radio is 1"],
+      sites: "1",
+      format: "json"
+    )
     assert_response :success
   end
 end